+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - added `change-grayscale` filter for change pixel values of grayscale images by a factor of fixed value
+- added `clip-grayscale` filter for replacing pixel values that go below or above thresholds
 
 
 0.0.1 (2024-05-06)